% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{Generic Power Function}
\usage{
delta(a = 0, b = 0, c = 0)
}
\arguments{
\item{a}{of \code{expression(ax^2 + bx + c )}}

\item{b}{of \code{expression(ax^2 + bx + c )}}

\item{c}{of \code{expression(ax^2 + bx + c )}}
}
\value{
The roots \code{x}, calculated by \code{a,b,c}.
}
\description{
This function raises a vector to any power. I'm making it internal,
by not \code{export}ing, because I don't want users to use it.
Why not? Because I said so!
}
