% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negtenvec.R
\docType{data}
\name{negtenvec}
\alias{negtenvec}
\title{Integer vector from -10 to -1}
\format{This (and the next) is a new tag that applies for data documentation.}
\source{
Note that you shouldn't use the `@export` tag when documenting data!
}
\usage{
negtenvec
}
\description{
Vector of (-1  -2  -3  -4  -5  -6  -7  -8  -9 -10)
}
\keyword{datasets}
