% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerPackage.R
\name{npower}
\alias{npower}
\title{power with n calculation}
\usage{
npower(x, n)
}
\arguments{
\item{x}{numeric}

\item{n}{numeric}
}
\value{
numeric
}
\description{
power with n calculation
}
\examples{
npower(2,5)
}
