% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerPackage.R
\docType{package}
\name{PowerPackage-package}
\alias{PowerPackage}
\alias{PowerPackage-package}
\title{PowerPackage: SSAI R WorkShop Exercise14}
\description{
squre,cube and power calculation
}
\author{
\strong{Maintainer}: Michael Su \email{michaelsuaustralia@hotmail.com}

}
\keyword{internal}
