% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrtc.R
\name{quadrtc}
\alias{quadrtc}
\title{Quadratic Formula}
\usage{
quadrtc(a, b, c, genplot = FALSE)
}
\arguments{
\item{genplot}{Generates and prints a plot of \code{x}
\code{FALSE} by default.}
}
\value{
The value(s) of x based on \code{x}
}
\description{
Solves for x for a quadratic equation: \code{expression(ax^2 + bx + c )}
Where x represents an unknown
While a, b, and c are constants with a not equal to 0
}
\details{
Verify that the quadratic formula satisfies the quadratic equation.
The root of the quadratic equation is returned.
For more information about the quadratic formula look
here \code{\link{https://en.wikipedia.org/wiki/Quadratic_formula}}
}
\examples{
quadrtc(2, -4, 2)
quadrtc(2, -3, 1)
quadrtc(-1, 0, -2)

}
