% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\alias{reciprocal}
\title{Square a vector}
\usage{
square(x)

reciprocal(x)
}
\arguments{
\item{x}{The vector to be squared.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item square (for \code{square})
     \item reciprocal/inverse (for \code{reciprocal})
}
 of \code{x}.
}
\description{
That's it -- these functions just square take the
reciprocal/inverse of a vector.
}
\details{
If you really want to see more, check out the internal \code{\link{pow}}
function that these functions depend on.
}
\examples{
square(tenvec)
square(-5)
reciprocal(2)
}
