% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrtc.R
\name{delta}
\alias{delta}
\title{Define internal function for delta
Rearrange expression(ax^2 + bx + c) to delta expression(delta = b^2 - 4ac)}
\usage{
delta(a, b, c)
}
\description{
Define internal function for delta
Rearrange expression(ax^2 + bx + c) to delta expression(delta = b^2 - 4ac)
}
